modify compile error

+++ kernel/kalloc.c
@@ -51,7 +51,7 @@ kfree(void *pa)
   if(((uint64)pa % PGSIZE) != 0 || (char*)pa < end || (uint64)pa >= PHYSTOP)
     panic("kfree");
 
-  int nthpage = pa / PGSIZE;
+  int nthpage =((uint64) pa) / PGSIZE;
   int refCount = derCowCount(nthpage);
   //cow_count[pa / PGSIZE]--;
   if (refCount != 0) {
